# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riogas Gestión API",
    "description": "Catálogo de APIs de GOYA (app 3). Backend NestJS (prefijo /api) + route handlers del front Next.\nGenerado por `pnpm docs:api` con un parser estático del código; no editar a mano — las descripciones y ejemplos van en docs/api/anotaciones.yaml.",
    "version": "0.1.0"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CrearLoteDto": {
        "type": "object",
        "properties": {
          "cantidad": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          }
        },
        "required": [
          "cantidad"
        ]
      },
...
      "CreateSorteoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500
          },
          "premioDescripcion": {
            "type": "string",
            "minLength": 3,
            "maxLength": 300
          },
          "fechaDesde": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHasta": {
            "type": "string",
            "format": "date-time",
            "description": "El orden (fechaHasta > fechaDesde) lo valida el service: en el PATCH parcial\nhay que compararlo contra los valores ya guardados."
          },
          "cantidadPremios": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "maxRegistrosDispositivoDia": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "edadMinima": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          }
        },
        "required": [
          "cantidadPremios",
          "fechaDesde",
          "fechaHasta",
          "nombre",
          "premioDescripcion"
        ]
      },
...
      "UpdateSorteoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500
          },
          "premioDescripcion": {
            "type": "string",
            "minLength": 3,
            "maxLength": 300
          },
          "fechaDesde": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHasta": {
            "type": "string",
            "format": "date-time"
          },
          "cantidadPremios": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "maxRegistrosDispositivoDia": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "edadMinima": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          }
        },
        "description": "Todos los campos opcionales (PATCH). El estado no se cambia acá: va por\nactivar / finalizar / cancelar, que tienen sus propias reglas de transición."
      },
...
    }
  },
  "paths": {
...
    "/api/sorteos": {
      "get": {
        "operationId": "SorteosAdminController_listar",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Búsqueda libre por nombre"
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Listado paginado de sorteos, con filtros.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "SorteosAdminController_crear",
        "tags": [
          "sorteos"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Alta de sorteo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSorteoDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/sorteos/{id}": {
      "get": {
        "operationId": "SorteosAdminController_detalle",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Detalle de un sorteo con sus lotes de códigos.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SorteosAdminController_actualizar",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Modificación parcial de un sorteo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSorteoDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/activar": {
      "post": {
        "operationId": "SorteosAdminController_activar",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Pone el sorteo en curso: empieza a aceptar participaciones.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/cancelar": {
      "post": {
        "operationId": "SorteosAdminController_cancelar",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Cancela el sorteo.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/finalizar": {
      "post": {
        "operationId": "SorteosAdminController_finalizar",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Cierra el sorteo: deja de aceptar participaciones.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/lotes": {
      "post": {
        "operationId": "SorteosAdminController_crearLote",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Genera un lote de códigos de participación para el sorteo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearLoteDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SorteosAdminController_listarLotes",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Lotes de códigos generados para el sorteo.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/lotes/{loteId}/zip": {
      "get": {
        "operationId": "SorteosAdminController_zipDelLote",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "loteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Descarga en streaming un ZIP con los PNG de los QR del lote.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/participaciones": {
      "get": {
        "operationId": "SorteosAdminController_listarParticipaciones",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Búsqueda libre por nombre / teléfono / email / código de canje"
          },
          {
            "name": "soloGanadores",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Se lee de `obj` y no de `value`: el enableImplicitConversion global ya\nconvirtió el string a boolean (y 'false' → true) antes de llegar acá."
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Listado paginado de participaciones del sorteo.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/{id}/participaciones/export": {
      "get": {
        "operationId": "SorteosAdminController_exportarParticipaciones",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Exporta en streaming las participaciones del sorteo a CSV.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/sorteos/participaciones/{id}/entregar": {
      "post": {
        "operationId": "SorteosAdminController_marcarPremioEntregado",
        "tags": [
          "sorteos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/sorteos/sorteos-admin.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Marca la participación ganadora como premio entregado.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
...
}